git_workplace/requirements.txt
joblib==1.1.0
numpy==1.22.0
pandas==1.3.5
scipy==1.7.3
scikit-learn==1.0.2
tensorflow-gpu==2.7.0
keras==2.7.0
tensorflow-addons==0.16.1